requirements.txt
apache-airflow==2.3.0  # Use the latest compatible version
celery
pendulum==2.1.2
psycopg2-binary
apache-airflow-providers-postgres
